feat: extract IssueActivityActionSentence component
